1.用户催单和接受订单接口已经完成，但是仍然是因为微信支付接口存在问题而无法测试

+++ sky-server/src/main/java/com/sky/controller/user/OrdersController.java
@@ -49,7 +49,7 @@ public class OrdersController {
 
     @GetMapping("/reminder/{id}")
     @ApiOperation("用户催单")
-    public Result reminder(@PathVariable Integer id) {
+    public Result reminder(@PathVariable Long id) {
         ordersService.reminder(id);
         return Result.success();
     }
